Fix two warnings on Win32.
* subversion/libsvn_subr/error.c (print_error): file_utf8 is only used
  when SVN_DEBUG is defined, so move the declaration there to avoid an
  unused-variable warning.

* subversion/svnserve/serve.c (serve): Explicitly cast the protocol
  version to an int. I presume the cast is safe, as I can't imagine a
  protocol with over 2 billion versions.


git-svn-id: https://svn.apache.org/repos/asf/subversion/trunk@850012 13f79535-47bb-0310-9956-ffa450edef68

+++ subversion/svnserve/serve.c
@@ -1082,7 +1082,7 @@ svn_error_t *serve(svn_ra_svn_conn_t *conn, serve_params_t *params,
   first = &APR_ARRAY_IDX(item->u.list, 0, svn_ra_svn_item_t);
   if (first->kind != SVN_RA_SVN_NUMBER)
     return SVN_NO_ERROR;
-  b.protocol_version = first->u.number;
+  b.protocol_version = (int) first->u.number;
   if (b.protocol_version == 1)
     {
       /* Version 1: auth exchange is mixed with client version and
